+++ source/core/aspx/encoder/base64.js
@@ -7,10 +7,12 @@
 'use strict';
 
 module.exports = (pwd, data, ext = null) => {
-  let randomID = `_0x${Math
-    .random()
-    .toString(16)
-    .substr(2)}`;
+  let randomID;
+  if (ext.opts.otherConf['use-random-variable'] === 1) {
+    randomID = antSword.utils.RandomChoice(antSword['RANDOMWORDS']);
+  } else {
+    randomID = `${antSword['utils'].RandomLowercase()}${Math.random().toString(16).substr(2)}`;
+  }
   data[randomID] = Buffer
     .from(data['_'])
     .toString('base64');